+++ nethsm/client/apis/path_to_api.py
@@ -1,12 +1,15 @@
 import typing
 import typing_extensions
 
+from nethsm.client.apis.paths.cluster_force_new import ClusterForceNew
 from nethsm.client.apis.paths.cluster_join import ClusterJoin
 from nethsm.client.apis.paths.cluster_members import ClusterMembers
 from nethsm.client.apis.paths.cluster_members_member_id import ClusterMembersMemberID
+from nethsm.client.apis.paths.cluster_members_member_id_promote import ClusterMembersMemberIDPromote
 from nethsm.client.apis.paths.config_backup_passphrase import ConfigBackupPassphrase
 from nethsm.client.apis.paths.config_logging import ConfigLogging
 from nethsm.client.apis.paths.config_network import ConfigNetwork
+from nethsm.client.apis.paths.config_ntp import ConfigNtp
 from nethsm.client.apis.paths.config_time import ConfigTime
 from nethsm.client.apis.paths.config_tls_cert_pem import ConfigTlsCertPem
 from nethsm.client.apis.paths.config_tls_cluster_ca_pem import ConfigTlsClusterCaPem
@@ -16,6 +19,7 @@ from nethsm.client.apis.paths.config_tls_public_pem import ConfigTlsPublicPem
 from nethsm.client.apis.paths.config_unattended_boot import ConfigUnattendedBoot
 from nethsm.client.apis.paths.config_unlock_passphrase import ConfigUnlockPassphrase
 from nethsm.client.apis.paths.health_alive import HealthAlive
+from nethsm.client.apis.paths.health_diagnose import HealthDiagnose
 from nethsm.client.apis.paths.health_ready import HealthReady
 from nethsm.client.apis.paths.health_state import HealthState
 from nethsm.client.apis.paths.info import Info
@@ -26,6 +30,7 @@ from nethsm.client.apis.paths.keys_key_id_cert import KeysKeyIDCert
 from nethsm.client.apis.paths.keys_key_id_csr_pem import KeysKeyIDCsrPem
 from nethsm.client.apis.paths.keys_key_id_decrypt import KeysKeyIDDecrypt
 from nethsm.client.apis.paths.keys_key_id_encrypt import KeysKeyIDEncrypt
+from nethsm.client.apis.paths.keys_key_id_label import KeysKeyIDLabel
 from nethsm.client.apis.paths.keys_key_id_move import KeysKeyIDMove
 from nethsm.client.apis.paths.keys_key_id_public_pem import KeysKeyIDPublicPem
 from nethsm.client.apis.paths.keys_key_id_restrictions_tags_tag import KeysKeyIDRestrictionsTagsTag
@@ -56,12 +61,15 @@ from nethsm.client.apis.paths.users_user_id_tags_tag import UsersUserIDTagsTag
 PathToApi = typing.TypedDict(
     'PathToApi',
     {
+    "/cluster/force-new": typing.Type[ClusterForceNew],
     "/cluster/join": typing.Type[ClusterJoin],
     "/cluster/members": typing.Type[ClusterMembers],
     "/cluster/members/{MemberID}": typing.Type[ClusterMembersMemberID],
+    "/cluster/members/{MemberID}/promote": typing.Type[ClusterMembersMemberIDPromote],
     "/config/backup-passphrase": typing.Type[ConfigBackupPassphrase],
     "/config/logging": typing.Type[ConfigLogging],
     "/config/network": typing.Type[ConfigNetwork],
+    "/config/ntp": typing.Type[ConfigNtp],
     "/config/time": typing.Type[ConfigTime],
     "/config/tls/cert.pem": typing.Type[ConfigTlsCertPem],
     "/config/tls/cluster-ca.pem": typing.Type[ConfigTlsClusterCaPem],
@@ -71,6 +79,7 @@ PathToApi = typing.TypedDict(
     "/config/unattended-boot": typing.Type[ConfigUnattendedBoot],
     "/config/unlock-passphrase": typing.Type[ConfigUnlockPassphrase],
     "/health/alive": typing.Type[HealthAlive],
+    "/health/diagnose": typing.Type[HealthDiagnose],
     "/health/ready": typing.Type[HealthReady],
     "/health/state": typing.Type[HealthState],
     "/info": typing.Type[Info],
@@ -81,6 +90,7 @@ PathToApi = typing.TypedDict(
     "/keys/{KeyID}/csr.pem": typing.Type[KeysKeyIDCsrPem],
     "/keys/{KeyID}/decrypt": typing.Type[KeysKeyIDDecrypt],
     "/keys/{KeyID}/encrypt": typing.Type[KeysKeyIDEncrypt],
+    "/keys/{KeyID}/label": typing.Type[KeysKeyIDLabel],
     "/keys/{KeyID}/move": typing.Type[KeysKeyIDMove],
     "/keys/{KeyID}/public.pem": typing.Type[KeysKeyIDPublicPem],
     "/keys/{KeyID}/restrictions/tags/{Tag}": typing.Type[KeysKeyIDRestrictionsTagsTag],
@@ -112,12 +122,15 @@ PathToApi = typing.TypedDict(
 
 path_to_api = PathToApi(
     {
+    "/cluster/force-new": ClusterForceNew,
     "/cluster/join": ClusterJoin,
     "/cluster/members": ClusterMembers,
     "/cluster/members/{MemberID}": ClusterMembersMemberID,
+    "/cluster/members/{MemberID}/promote": ClusterMembersMemberIDPromote,
     "/config/backup-passphrase": ConfigBackupPassphrase,
     "/config/logging": ConfigLogging,
     "/config/network": ConfigNetwork,
+    "/config/ntp": ConfigNtp,
     "/config/time": ConfigTime,
     "/config/tls/cert.pem": ConfigTlsCertPem,
     "/config/tls/cluster-ca.pem": ConfigTlsClusterCaPem,
@@ -127,6 +140,7 @@ path_to_api = PathToApi(
     "/config/unattended-boot": ConfigUnattendedBoot,
     "/config/unlock-passphrase": ConfigUnlockPassphrase,
     "/health/alive": HealthAlive,
+    "/health/diagnose": HealthDiagnose,
     "/health/ready": HealthReady,
     "/health/state": HealthState,
     "/info": Info,
@@ -137,6 +151,7 @@ path_to_api = PathToApi(
     "/keys/{KeyID}/csr.pem": KeysKeyIDCsrPem,
     "/keys/{KeyID}/decrypt": KeysKeyIDDecrypt,
     "/keys/{KeyID}/encrypt": KeysKeyIDEncrypt,
+    "/keys/{KeyID}/label": KeysKeyIDLabel,
     "/keys/{KeyID}/move": KeysKeyIDMove,
     "/keys/{KeyID}/public.pem": KeysKeyIDPublicPem,
     "/keys/{KeyID}/restrictions/tags/{Tag}": KeysKeyIDRestrictionsTagsTag,