+++ nethsm/client/components/schema/key_generate_request_data.py
@@ -24,6 +24,7 @@ class Length(
     inclusive_minimum: typing.Union[int, float] = 128
 
 from nethsm.client.components.schema import id
+from nethsm.client.components.schema import key_label
 from nethsm.client.components.schema import key_mechanisms
 from nethsm.client.components.schema import key_restrictions
 from nethsm.client.components.schema import key_type
@@ -35,6 +36,7 @@ Properties = typing.TypedDict(
         "length": typing.Type[Length],
         "id": typing.Type[id.ID],
         "restrictions": typing.Type[key_restrictions.KeyRestrictions],
+        "label": typing.Type[key_label.KeyLabel],
     }
 )
 
@@ -49,6 +51,7 @@ class KeyGenerateRequestDataDict(schemas.immutabledict[str, schemas.OUTPUT_BASE_
         "length",
         "id",
         "restrictions",
+        "label",
     })
     
     def __new__(
@@ -83,6 +86,10 @@ class KeyGenerateRequestDataDict(schemas.immutabledict[str, schemas.OUTPUT_BASE_
             key_restrictions.KeyRestrictionsDict,
             schemas.Unset
         ] = schemas.unset,
+        label: typing.Union[
+            str,
+            schemas.Unset
+        ] = schemas.unset,
         configuration_: typing.Optional[schema_configuration.SchemaConfiguration] = None,
         **kwargs: schemas.INPUT_TYPES_ALL,
     ):
@@ -94,6 +101,7 @@ class KeyGenerateRequestDataDict(schemas.immutabledict[str, schemas.OUTPUT_BASE_
             ("length", length),
             ("id", id),
             ("restrictions", restrictions),
+            ("label", label),
         ):
             if isinstance(val, schemas.Unset):
                 continue
@@ -153,6 +161,16 @@ class KeyGenerateRequestDataDict(schemas.immutabledict[str, schemas.OUTPUT_BASE_
             return val
         return typing.cast(
             key_restrictions.KeyRestrictionsDict,
+            val
+        )
+    
+    @property
+    def label(self) -> typing.Union[str, schemas.Unset]:
+        val = self.get("label", schemas.unset)
+        if isinstance(val, schemas.Unset):
+            return val
+        return typing.cast(
+            str,
             val
         )