+++ nethsm/client/components/schema/private_key_pem.py
@@ -11,6 +11,7 @@ from __future__ import annotations
 from nethsm.client.shared_imports.schema_imports import *  # pyright: ignore [reportWildcardImportFromLibrary]
 
 
+from nethsm.client.components.schema import key_label
 from nethsm.client.components.schema import key_mechanisms
 from nethsm.client.components.schema import key_restrictions
 from nethsm.client.components.schema import pem_private_key
@@ -19,6 +20,7 @@ Properties = typing.TypedDict(
     {
         "mechanisms": typing.Type[key_mechanisms.KeyMechanisms],
         "restrictions": typing.Type[key_restrictions.KeyRestrictions],
+        "label": typing.Type[key_label.KeyLabel],
     }
 )
 
@@ -30,6 +32,7 @@ class ArgumentsDict(schemas.immutabledict[str, schemas.OUTPUT_BASE_TYPES]):
     __optional_keys__: typing.FrozenSet[str] = frozenset({
         "mechanisms",
         "restrictions",
+        "label",
     })
     
     def __new__(
@@ -45,6 +48,10 @@ class ArgumentsDict(schemas.immutabledict[str, schemas.OUTPUT_BASE_TYPES]):
             key_restrictions.KeyRestrictionsDict,
             schemas.Unset
         ] = schemas.unset,
+        label: typing.Union[
+            str,
+            schemas.Unset
+        ] = schemas.unset,
         configuration_: typing.Optional[schema_configuration.SchemaConfiguration] = None,
         **kwargs: schemas.INPUT_TYPES_ALL,
     ):
@@ -52,6 +59,7 @@ class ArgumentsDict(schemas.immutabledict[str, schemas.OUTPUT_BASE_TYPES]):
         for key_, val in (
             ("mechanisms", mechanisms),
             ("restrictions", restrictions),
+            ("label", label),
         ):
             if isinstance(val, schemas.Unset):
                 continue
@@ -87,6 +95,16 @@ class ArgumentsDict(schemas.immutabledict[str, schemas.OUTPUT_BASE_TYPES]):
             return val
         return typing.cast(
             key_restrictions.KeyRestrictionsDict,
+            val
+        )
+    
+    @property
+    def label(self) -> typing.Union[str, schemas.Unset]:
+        val = self.get("label", schemas.unset)
+        if isinstance(val, schemas.Unset):
+            return val
+        return typing.cast(
+            str,
             val
         )