+++ nethsm/client/paths/keys_key_prefix/get/operation.py
@@ -15,13 +15,17 @@ from .responses import (
     response_406,
     response_412,
 )
-from .parameters import parameter_0
+from .parameters import (
+    parameter_0,
+    parameter_1,
+)
 from ..parameters import parameter_0 as path_item_parameter_0
 from .security import security_requirement_object_0
 from .path_parameters import PathParameters, PathParametersDictInput, PathParametersDict
 from .query_parameters import QueryParameters, QueryParametersDictInput, QueryParametersDict
 query_parameter_classes = (
     parameter_0.Parameter0,
+    parameter_1.Parameter1,
 )
 path_parameter_classes = (
     path_item_parameter_0.Parameter0,